Fix repo updates, app upgrades

+++ SharpPackageManager/main.cs
@@ -20,7 +20,7 @@ public class SharpPackageManager
     public static Dictionary<string, string> repos = new Dictionary<string, string>();
     public static void Main(string[] args)
     {
-        if (!System.IO.Directory.Exists("C:\\SPM\\futureversion") & System.IO.File.Exists("C:\\SPM\\futureversion\\unlock.txt")) 
+        if (System.IO.Directory.Exists("C:\\SPM\\futureversion") && !System.IO.File.Exists("C:\\SPM\\futureversion\\unlock.txt"))
         {
             Console.WriteLine("Unlocking update on app start and executing the app...");
             System.IO.File.Create("C:\\SPM\\futureversion\\unlock.txt");
@@ -30,7 +30,7 @@ public class SharpPackageManager
             PackageStartInfo.StartInfo.Verb = "runas";
             PackageStartInfo.Start();
         }
-        else if (System.IO.File.Exists("C:\\SPM\\futureversion\\unlock.txt"))
+        else if (System.IO.File.Exists("C:\\SPM\\futureversion\\unlock.txt") && System.IO.Directory.Exists("C:\\SPM\\futureversion"))
         {
             Console.WriteLine("Update is unlocked, starting the main upgrade script...");
             System.IO.File.Delete("C:\\SPM\\futureversion\\unlock.txt");
@@ -38,49 +38,65 @@ public class SharpPackageManager
             string[] upfiles = System.IO.Directory.GetFiles("C:\\SPM\\futureversion\\SPM");
             foreach (string upfile in upfiles)
             {
-                string fileName= System.IO.Path.GetFileName(upfile);
+                string fileName = System.IO.Path.GetFileName(upfile);
                 string destFile = System.IO.Path.Combine(InstallPath + fileName);
                 System.IO.File.Copy(upfile, destFile, true);
 
             }
+            System.IO.File.Create(InstallDir + "clean.txt");
+            Console.WriteLine("Please restart the app!");
+        }
+        else if (System.IO.File.Exists("C:\\SPM\\clean.txt"))
+        {
             Console.WriteLine("Cleaning Up...");
             System.IO.Directory.Delete(InstallPath + "futureversion", true);
             System.IO.File.Delete("C:\\SPM.zip");
+            System.IO.File.Delete(InstallDir + "clean.txt");
             Console.WriteLine("Update Finished!");
-
         }
         else
         {
-            DataLoad(InstallDir + "sources.txt", "repos");
-            DataLoad(InstallDir + "apps.txt", "apps");
-            Console.WriteLine("Please choose your action! \n \n");
-            Console.WriteLine("Install a package (Command: i) \n \n");
-            Console.WriteLine("Install an AppKit (Command: ak) \n \n");
-            Console.WriteLine("Update database (Command: up) \n \n");
-            Console.WriteLine("Check for SPM updates (Command: spmup)");
-            string action = Console.ReadLine();
-            if (action == "i") // | action == "install")
-            {
-                Console.WriteLine("Which Package do you want to install?");
-                //string Package = Console.ReadLine();
-                InstallPkg(Console.ReadLine());
-            }
-            else if (action == "up")
-            {
-                DataUpdate(InstallDir + "apps.txt", "apps");
-            } //| action == "update");
-            else if (action == "ak")
-            {
-                Console.WriteLine("Please write the appkit txt file path (also please enter path with double \\ (example C:\\\\example\\\\appkit.txt)");
-                string kitpath = (Console.ReadLine());
-                if (kitpath != null) AppKits(kitpath);
-                else Console.WriteLine("Kit path has to be something another (not null)");
-            }
-            else if (action == "spmup") VersionUpdate(curbranch);
+            MainApp();
         }
-        //InstallPkg("steam");
+       
     }
-    public static void VersionUpdate(string branch)
+    public static void MainApp()
+    {
+        if (!System.IO.Directory.Exists("C:\\SPM\\Downloads")) System.IO.Directory.CreateDirectory("C:\\SPM\\Downloads");
+        if (!System.IO.Directory.Exists("C:\\SPM\\config"))
+        {
+            System.IO.Directory.CreateDirectory("C:\\SPM\\config");
+
+        }
+
+        DataLoad(InstallDir + "sources.txt", "repos");
+        //DataLoad(InstallDir + "apps.txt", "apps");
+        Console.WriteLine("Please choose your action! \n \n");
+        Console.WriteLine("Install a package (Command: i) \n \n");
+        Console.WriteLine("Install an AppKit (Command: ak) \n \n");
+        Console.WriteLine("Update database (Command: up) \n \n");
+        Console.WriteLine("Check for SPM updates (Command: spmup)");
+        string action = Console.ReadLine();
+        if (action == "i") // | action == "install")
+        {
+            Console.WriteLine("Which Package do you want to install?");
+            //string Package = Console.ReadLine();
+            InstallPkg(Console.ReadLine());
+        }
+        else if (action == "up")
+        {
+            DataUpdate();
+        } //| action == "update");
+        else if (action == "ak")
+        {
+            Console.WriteLine("Please write the appkit txt file path (also please enter path with double \\ (example C:\\\\example\\\\appkit.txt)");
+            string kitpath = (Console.ReadLine());
+            if (kitpath != null) AppKits(kitpath);
+            else Console.WriteLine("Kit path has to be something another (not null)");
+        }
+        else if (action == "spmup") VersionUpdate(curbranch);
+    }
+        public static void VersionUpdate(string branch)
     {
         Console.WriteLine("Loading latest versions info...");
         using (WebClient tagdl = new WebClient())
@@ -171,28 +187,27 @@ public class SharpPackageManager
         }
         else Console.WriteLine("Please specify the package correctly!");
     }
-    public static void DataUpdate(string File, string Type)
+    public static void DataUpdate()
     {
-
-        {
+            appnames.Clear();
+            appurls.Clear();
+            repos.Clear();
             using (WebClient srcdl = new WebClient())
             {
                 int i = 0;
                 do
                 {
-                    
+                    string currepopath = InstallDir + "apps" + reponames[i] + ".txt";
+                    Console.WriteLine(currepopath);
                     Console.WriteLine("Updating "+reponames[i]);
-                    srcdl.DownloadFile(repourls[i] + "/apps.txt", InstallDir+"apps.txt");
+                    srcdl.DownloadFile(repourls[i] + "/apps.txt", InstallDir+"apps"+reponames[i]+".txt");
                     i++;
+                    DataLoad(currepopath, "apps");
                     // Param1 = Link of file
                     // Param2 = Path to save
                 } while (i != repourls.Count());
             }
-            appnames.Clear();
-            appurls.Clear();
-            repos.Clear();
-            DataLoad(InstallDir+"apps.txt", "apps");
-        }
+            MainApp();
         }
     public static void DataLoad(string File, string Type)
     {
